+++ test/built-imports.ts
@@ -1,5 +1,5 @@
 import t from 'tap'
-import getImports from '../src/get-imports.js'
+import getImports from '../src/built-imports.js'
 import { Package } from '../src/types.js'
 
 const cases: [
